[FIX] Use i64 instead of i32 for Ids..

+++ migrations/2018-03-24-170906_create_wallets/up.sql
@@ -1,10 +1,10 @@
 CREATE TABLE wallets (
-    id              SERIAL NOT NULL PRIMARY KEY,
-    user_id         INTEGER NOT NULL REFERENCES users(id),
+    id              BIGSERIAL NOT NULL PRIMARY KEY,
+    user_id         BIGINT NOT NULL REFERENCES users(id),
     name            TEXT NOT NULL,
     wallet_type     TEXT NOT NULL,
-    initial_balance INTEGER NOT NULL,
-    current_balance INTEGER NOT NULL,
+    initial_balance BIGINT NOT NULL,
+    current_balance BIGINT NOT NULL,
     color           TEXT NOT NULL,
     created_at      TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 )
